+++ src/ocr_app/data_process/data_model/ocr_output.py
@@ -1,19 +1,12 @@
-import os
 from dataclasses import dataclass
 from typing import Dict, List
 
 import numpy as np
 
-FILE_PATH = os.path.join(
-    os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(__file__))))),
-    'results',
-)
-print(FILE_PATH)
-if not os.path.exists(FILE_PATH):
-    os.makedirs(FILE_PATH)
-
 
 @dataclass
 class OcrResults:
+    """Class for keeping track of OCR model output."""
+
     images: List[np.ndarray]
     ocr_outputs: List[List[Dict[str, Dict[str, int]]]]
