+++ src/ocr_app/data_process/output_handler/image_process.py
